+++ oci/core_virtual_circuit_public_prefix_test.go
@@ -16,7 +16,8 @@ var (
 		"verification_state": Representation{repType: Optional, create: `COMPLETED`},
 	}
 
-	VirtualCircuitPublicPrefixResourceConfig = VirtualCircuitPublicPropertyVariables + VirtualCircuitPublicRequiredOnlyResource
+	VirtualCircuitPublicPrefixResourceConfig = VirtualCircuitPublicPropertyVariables + VirtualCircuitResourceDependencies +
+		generateResourceFromRepresentationMap("oci_core_virtual_circuit", "test_virtual_circuit", Required, Create, virtualCircuitPublicRequiredOnlyRepresentation)
 )
 
 func TestCoreVirtualCircuitPublicPrefixResource_basic(t *testing.T) {